+++ deploy/07_default_strategy.ts
@@ -1,11 +1,6 @@
 import { DeployFunction } from "@celo/staked-celo-hardhat-deploy/types";
-import { BigNumber } from "ethers";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { catchNotOwnerForProxy } from "../lib/deploy-utils";
-import { ADDRESS_ZERO } from "../test-ts/utils";
-
-const parseValidatorGroups = (validatorGroupsString: string | undefined) =>
-  validatorGroupsString ? validatorGroupsString.split(",") : [];
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deploy } = hre.deployments;
@@ -15,8 +10,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const lib = await hre.ethers.getContractFactory("AddressSortedLinkedList");
   const libInstance = await lib.deploy();
   await libInstance.deployed();
-
-  const isDefaultStrategyAlreadyDeployed = await hre.deployments.getOrNull("DefaultStrategy");
 
   const managerAddress = (await hre.deployments.get("Manager")).address;
   await catchNotOwnerForProxy(
@@ -36,38 +29,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
       },
     })
   );
-
-  if (!isDefaultStrategyAlreadyDeployed) {
-    const validatorGroups = parseValidatorGroups(process.env.VALIDATOR_GROUPS);
-    if (validatorGroups.length == 0) {
-      return;
-    }
-
-    const defaultStrategy = await hre.ethers.getContract("DefaultStrategy");
-    const account = await hre.ethers.getContract("Account");
-
-    const validatorGroupsWithCelo = await Promise.all(
-      validatorGroups.map(async (validatorGroup) => ({
-        group: validatorGroup,
-        celo: (await account.getCeloForGroup(validatorGroup)) as BigNumber,
-      }))
-    );
-
-    const validatorGroupsSortedDesc = validatorGroupsWithCelo.sort((a, b) =>
-      a.celo.lt(b.celo) ? 1 : -1
-    );
-
-    let nextGroup = ADDRESS_ZERO;
-    for (let i = 0; i < validatorGroupsSortedDesc.length; i++) {
-      const activateTx = await defaultStrategy.activateGroup(
-        validatorGroupsSortedDesc[i].group,
-        ADDRESS_ZERO,
-        nextGroup
-      );
-      await activateTx.wait();
-      nextGroup = validatorGroupsSortedDesc[i].group;
-    }
-  }
 };
 
 func.id = "deploy_default_strategy";
